2Reasoning_LLM/requirements.txt
azure-core
azure-ai-inference
python-dotenv
flask
uvicorn